Updated RunnableExample class
Hopefully I committed this right.
At first, I wanted to use the BayesClassifier class to do all of the
work for me, just to see if it could. I parsed each tweet and had the
bayes object "learn" the tweets for each category (APP and OTHER). I
don't know if I did something wrong or the BayesClassifier algorithm is
a little different, but the results were slightly off for a couple of
testTweets.

So I decided to make the algorithm myself, converting the DataSmart
excel code into Java code. It worked! Now, I believe the parsing was a
tiny bit off so the exact number of tokens differed a tiny bit. Compared
to the excel code, my code generated 19/20 of the same outputs (with the
only difference being a very slight amount, probably due to the
parsing).