RunnableExample Class
I believe I committed the wrong file.
Parsed the APP and OTHER csv files, and made the "bayes" object learn
each tweet for its respective category.  Then parsed the testTweets and
used the bayes object to classify each test tweet. The results were
pretty mixed compared to the DataSmart excel results.

I then converted the DataSmart excel code into Java code.  I calculated
the probability for each token in the APP and OTHER categories. I then
Parsed the testTweets and calculated the probability of the tweet being
about the APP or about OTHER. The results went smoothly, getting 19/20
testTweets the same as the excel file.
I believe the parsing of the tweets was a little off in my java code
compared to the excel document, leading to differing tokens and token
counts.